v0.4-Raw-A2A/requirements.txt
mcp[cli]>=1.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
a2a-sdk>=0.2.0
httpx>=0.27.0
uvicorn>=0.29.0
fastapi>=0.111.0